requirements.txt
pandas==1.3.5
python-dotenv==0.20.0
requests==2.27.1
